Optimized components and added some styling to site.
