+++ src/App.css
@@ -1,0 +1,41 @@
+/* . {
+    margin: 0px;
+    padding: 0px;
+} */
+
+h1 {
+    background: linear-gradient(to right, black 0%,rgb(158, 158, 158) 25%,rgb(158, 158, 158) 76%,black 100%);
+    height: 4rem;
+    text-align: center;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 0px;
+    margin-left: 0px;
+    margin: 0px;
+    padding-left: 1rem;
+    border-bottom-left-radius: 1rem;
+}
+
+input {
+    margin-top: 1rem;
+    /* height: 25vh;
+    width: 50vw; */
+}
+
+button {
+    margin-top: 1rem;
+    cursor: pointer;
+}
+
+.main-container {
+    height:100vh;
+    background: linear-gradient(0deg,rgb(253, 253, 151) 0%, rgb(255, 255, 85) 50% , rgb(241, 241, 95) 100% );
+}
+
+@media screen and (max-width: 600px) {
+    h1 {
+        background: linear-gradient(to right, black 0%,rgb(158, 158, 158) 25%,rgb(158, 158, 158) 76%,black 100%);
+        justify-content: center;
+    }
+}
