+++ src/Components/Card.js
@@ -2,12 +2,18 @@ import React from "react";
 import AudioControl from "./AudioControl";
 import DefinitionCarousel from "./DefinitionCarousel"
 import GeneralSymAnt from "./GeneralSymAnt";
+import { MdOutlinePlayArrow } from "react-icons/md";
 
 export default function Card({word,phonetic,audio,meanings,sourceURLs}) {
     return (
             <div>
                 <div>
-                    <h2>{word}</h2>
+                    <h2>
+                        {word} 
+                        <span>
+                            <MdOutlinePlayArrow />
+                        </span> 
+                    </h2> 
                     <h3>{phonetic}</h3>
                 </div>
                 <AudioControl audio={audio} />
